Moved OculusSimpleFireFlags into SimpleFireMessage
