+++ include/liboculus/SonarConfiguration.h
@@ -37,44 +37,6 @@
 
 namespace liboculus {
 
-struct FlagBits {
-  // bit 0: 0 = interpret range as percent, 1 = interpret range as meters
-  // bit 1: 0 = 8 bit data, 1 = 16 bit data
-  // bit 2: 0 = wont send gain, 1 = send gain
-  // bit 3: 0 = send full return message, 1 = send simple return message
-  // bit 4: "gain assistance"?
-  // bit 6: use 512 beams (vs 256): email from Blueprint said to set flags |= 0x40
-
-  static const uint8_t RangeAsMeters = (0x01) << 0;
-  static const uint8_t Data16Bits    = (0x01) << 1;
-  static const uint8_t DoSendGain    = (0x01) << 2;
-  static const uint8_t SimpleReturn  = (0x01) << 3;
-  static const uint8_t GainAssistance = (0x01) << 4;
-  static const uint8_t Do512Beams    = (0x01) << 6;
-};
-
-// This class used to be both a way to store state and
-// deserializer.  But it got awkward, so now it's 
-// *only* a parser of flag bytes (does no allow setting)
-// configuration is stored outside of the bitfield in
-// SonarConfiguration, and the flag byte is made just-in-time
-class OculusSimpleFireFlags {
- public:
-    OculusSimpleFireFlags() = delete;
-    OculusSimpleFireFlags( const uint8_t flags )
-    : _data(flags) {;}
-
-    bool getRangeAsMeters() const           { return _data & FlagBits::RangeAsMeters; }
-    bool getSendGain() const                { return _data & FlagBits::DoSendGain; }
-    bool getData16Bit() const               { return _data & FlagBits::Data16Bits; }
-    bool getSimpleReturn() const            { return _data & FlagBits::SimpleReturn; }
-    bool getGainAssistance() const          { return _data & FlagBits::GainAssistance; }
-    bool get512Beams() const                { return _data & FlagBits::Do512Beams;}
-
- private:
-    uint8_t _data;
-};
-
 // Thin OO wrapper around the OculusSimpleFireMessage.
 //
 // \todo This API is a little messy right now.  Probably *shouldn't* keep
